+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 FEATURES:
 * **New Resource:** sumologic_cse_entity_entity_group_configuration (GH-376)
 * **New Resource:** sumologic_cse_inventory_entity_group_configuration (GH-376)
+* Add new optional `notification_group_fields` field to resource/sumologic_monitor (GH-390)
 
 ## 2.16.2 (June 12, 2022)