+++ CHANGELOG.md
@@ -3,6 +3,10 @@ FEATURES:
 * **New Resource:** sumologic_cse_entity_entity_group_configuration (GH-376)
 * **New Resource:** sumologic_cse_inventory_entity_group_configuration (GH-376)
 
+BUG FIXES:
+* Allow Monitor move between Monitor folders
+
+
 ## 2.16.2 (June 12, 2022)
 
 BUG FIXES: